Before moving all packages to packages folder
