+++ .gitignore
@@ -1,24 +1,22 @@
-# Logs
-logs
-*.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-pnpm-debug.log*
-lerna-debug.log*
+# Folders
+dist/
+build/
+node_modules/
 
-node_modules
-dist
-dist-ssr
-*.local
-
-# Editor directories and files
-.vscode/*
-!.vscode/extensions.json
-.idea
+# File managers
 .DS_Store
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
-*.sw?
+.directory
+desktop.ini
+
+# Other lockfiles
+package-lock.json
+npm-shrinkwrap.json
+
+# Log files
+*.log*
+
+# Excel files
+*.xlsx
+
+# Now this is a typescript project
+src/index.js
