+++ example-express/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "example-express",
+  "version": "1.0.0",
+  "license": "UNLICENSED",
+  "scripts": {
+    "dev": "nodemon --exec ts-node src/server.ts"
+  },
+  "dependencies": {
+    "express": "^4.18.2",
+    "typescript": "^4.9.3"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.14",
+    "@types/node": "^18.11.11",
+    "nodemon": "^2.0.20",
+    "ts-node": "^10.9.1"
+  }
+}
