--- old/.gitignore
@@ -1,21 +1,0 @@
-# Folders
-dist/
-node_modules/
-
-# File managers
-.DS_Store
-.directory
-desktop.ini
-
-# Other lockfiles
-yarn.lock
-npm-shrinkwrap.json
-
-# Log files
-*.log*
-
-# Excel files
-*.xlsx
-
-# Now this is a typescript project
-src/index.js
