+++ old/reactjs-app/src/vite-env.d.ts
