+++ old/reactjs-app/tsconfig.json
