+++ old/reactjs-app/vite.config.ts
