+++ package.json
@@ -1,15 +1,10 @@
 {
-  "name": "json-as-xlsx",
   "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
-  },
-  "devDependencies": {
-    "typescript": "^4.6.4",
-    "vite": "^3.2.3"
+  "workspaces": [
+    "json-as-xlsx",
+    "example-express"
+  ],
+  "dependencies": {
+    "lerna": "^6.1.0"
   }
 }
